♻️ Move slog.go to opts and add alias for tests
