+++ level.go
@@ -1,8 +1,6 @@
 package sqlslog
 
 import (
-	"log/slog"
-
 	"github.com/akm/sql-slog/opts"
 )
 
@@ -16,13 +14,3 @@ const (
 	LevelWarn    Level = opts.LevelWarn
 	LevelError   Level = opts.LevelError
 )
-
-// ReplaceLevelAttr replaces the log level as sqlslog.Level with its string representation.
-func ReplaceLevelAttr(_ []string, a slog.Attr) slog.Attr {
-	// https://go.dev/src/log/slog/example_custom_levels_test.go
-	if a.Key == slog.LevelKey {
-		level := opts.Level(a.Value.Any().(slog.Level)) //nolint:forcetypeassert
-		a.Value = slog.StringValue(level.String())
-	}
-	return a
-}
